所有添加，第一次不添加图片，编辑时添加报错

+++ thinkphp/application/index/model/Common.php
@@ -30,9 +30,10 @@ class Common
     // 删除指定文件夹下的图片
     public static function deleteImage($imagePath)
     {
-        if(file_exists($imagePath)){
-            unlink($imagePath);
+        if(!is_null($imagePath)) {
+            if(file_exists($imagePath)){
+                unlink($imagePath);
+            }
         }
-        
     }
 }
